Moves sleep implementation to sleep_hal.cpp and stub the sleep functions in core HAL.
