+++ hal/src/nRF52840/core_hal.c
@@ -57,7 +57,7 @@
 #include "flash_common.h"
 #include <nrf_pwm.h>
 #include "concurrent_hal.h"
-#include "ble_hal.h"
+#include "sleep_hal.h"
 
 #define BACKUP_REGISTER_NUM        10
 static int32_t backup_register[BACKUP_REGISTER_NUM] __attribute__((section(".backup_registers")));
@@ -72,12 +72,6 @@ typedef enum Feature_Flag {
     FEATURE_FLAG_RESET_INFO = 0x01,
     FEATURE_FLAG_ETHERNET_DETECTION = 0x02
 } Feature_Flag;
-
-typedef enum {
-    STOP_MODE_EXIT_CONDITION_NONE = 0x00,
-    STOP_MODE_EXIT_CONDITION_PIN  = 0x01,
-    STOP_MODE_EXIT_CONDITION_RTC  = 0x02
-} stop_mode_exit_condition_t;
 
 void HardFault_Handler( void ) __attribute__( ( naked ) );
 
@@ -487,639 +481,30 @@ void HAL_Core_Enter_Bootloader(bool persist) {
 }
 
 void HAL_Core_Enter_Stop_Mode(uint16_t wakeUpPin, uint16_t edgeTriggerMode, long seconds) {
-    InterruptMode m = (InterruptMode)edgeTriggerMode;
-    HAL_Core_Enter_Stop_Mode_Ext(&wakeUpPin, 1, &m, 1, seconds, NULL);
-}
-
-static void fpu_sleep_prepare(void) {
-    uint32_t fpscr;
-    fpscr = __get_FPSCR();
-    /*
-     * Clear FPU exceptions.
-     * Without this step, the FPU interrupt is marked as pending,
-     * preventing system from sleeping. Exceptions cleared:
-     * - IOC - Invalid Operation cumulative exception bit.
-     * - DZC - Division by Zero cumulative exception bit.
-     * - OFC - Overflow cumulative exception bit.
-     * - UFC - Underflow cumulative exception bit.
-     * - IXC - Inexact cumulative exception bit.
-     * - IDC - Input Denormal cumulative exception bit.
-     */
-    __set_FPSCR(fpscr & ~0x9Fu);
-    __DMB();
-    NVIC_ClearPendingIRQ(FPU_IRQn);
-
-    /*__
-     * Assert no critical FPU exception is signaled:
-     * - IOC - Invalid Operation cumulative exception bit.
-     * - DZC - Division by Zero cumulative exception bit.
-     * - OFC - Overflow cumulative exception bit.
-     */
-    SPARK_ASSERT((fpscr & 0x07) == 0);
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
 }
 
 int HAL_Core_Enter_Stop_Mode_Ext(const uint16_t* pins, size_t pins_count, const InterruptMode* mode, size_t mode_count, long seconds, void* reserved) {
-    // Initial sanity check
-    if ((pins_count == 0 || mode_count == 0 || pins == NULL || mode == NULL) && seconds <= 0 && reserved == NULL) {
-        return SYSTEM_ERROR_NOT_ALLOWED;
-    }
-
-    // Validate pins and modes
-    if ((pins_count > 0 && pins == NULL) || (pins_count > 0 && mode_count == 0) || (mode_count > 0 && mode == NULL)) {
-        return SYSTEM_ERROR_NOT_ALLOWED;
-    }
-
-    for (unsigned i = 0; i < pins_count; i++) {
-        if (pins[i] >= TOTAL_PINS) {
-            return SYSTEM_ERROR_NOT_ALLOWED;
-        }
-    }
-
-    for (int i = 0; i < mode_count; i++) {
-        switch(mode[i]) {
-            case RISING:
-            case FALLING:
-            case CHANGE:
-                break;
-            default:
-                return SYSTEM_ERROR_NOT_ALLOWED;
-        }
-    }
-
-    bool wakeup_by_ble = false;
-    if (reserved) {
-        uint32_t flag = *((uint32_t*)reserved);
-        if (flag & 0x01) {
-            wakeup_by_ble = true;
-        }
-    }
-
-    // Detach USB
-    HAL_USB_Detach();
-
-    // Flush all USARTs
-    for (int usart = 0; usart < TOTAL_USARTS; usart++) {
-        if (HAL_USART_Is_Enabled(usart)) {
-            HAL_USART_Flush_Data(usart);
-        }
-    }
-
-    // Make sure we acquire exflash lock BEFORE going into a critical section
-    hal_exflash_lock();
-
-    // Disable thread scheduling
-    os_thread_scheduling(false, NULL);
-
-    // This will disable all but SoftDevice interrupts (by modifying NVIC->ICER)
-    uint8_t st = 0;
-    sd_nvic_critical_region_enter(&st);
-
-    // Disable SysTick
-    SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
-
-    // Put external flash into sleep and disable QSPI peripheral
-    hal_exflash_special_command(HAL_EXFLASH_COMMAND_NONE, HAL_EXFLASH_COMMAND_SLEEP, NULL, NULL, 0);
-    hal_exflash_uninit();
-
-    uint32_t exit_conditions = STOP_MODE_EXIT_CONDITION_NONE;
-
-    // Reducing power consumption
-    // Suspend all PWM instance remembering their state
-    bool pwm_state[4] = {
-        NRF_PWM0->ENABLE & PWM_ENABLE_ENABLE_Msk,
-        NRF_PWM1->ENABLE & PWM_ENABLE_ENABLE_Msk,
-        NRF_PWM2->ENABLE & PWM_ENABLE_ENABLE_Msk,
-        NRF_PWM3->ENABLE & PWM_ENABLE_ENABLE_Msk,
-    };
-    nrf_pwm_disable(NRF_PWM0);
-    nrf_pwm_disable(NRF_PWM1);
-    nrf_pwm_disable(NRF_PWM2);
-    nrf_pwm_disable(NRF_PWM3);
-
-    // _Attempt_ to disable HFCLK. This may not succeed, resulting in a higher current consumption
-    // FIXME
-    bool hfclk_resume = false;
-    if (nrf_drv_clock_hfclk_is_running()) {
-        hfclk_resume = true;
-
-        // Temporarily enable SoftDevice API interrupts just in case
-        uint32_t base_pri = __get_BASEPRI();
-        // We are also allowing IRQs with priority = 5, because that's what
-        // OpenThread uses for its SWI3
-        __set_BASEPRI(_PRIO_APP_LOW << (8 - __NVIC_PRIO_BITS));
-        sd_nvic_critical_region_exit(st);
-        {
-            nrf_drv_clock_hfclk_release();
-            // while (nrf_drv_clock_hfclk_is_running()) {
-            //     ;
-            // }
-        }
-        // And disable again
-        sd_nvic_critical_region_enter(&st);
-        __set_BASEPRI(base_pri);
-    }
-
-    // Remember current microsecond counter
-    uint64_t micros_before_sleep = hal_timer_micros(NULL);
-    // Disable hal_timer (RTC2)
-    hal_timer_deinit(NULL);
-
-    // Make sure LFCLK is running
-    nrf_drv_clock_lfclk_request(NULL);
-    while (!nrf_drv_clock_lfclk_is_running()) {
-        ;
-    }
-
-    // Configure RTC2 to count at 125ms interval. This should allow us to sleep
-    // (2^24 - 1) * 125ms = 24 days
-    NVIC_SetPriority(RTC2_IRQn, 4);
-    // Make sure that the RTC is stopped and cleared
-    nrf_rtc_task_trigger(NRF_RTC2, NRF_RTC_TASK_STOP);
-    nrf_rtc_task_trigger(NRF_RTC2, NRF_RTC_TASK_CLEAR);
-    __DSB();
-    __ISB();
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_TICK_MASK);
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_OVERFLOW_MASK);
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_COMPARE0_MASK);
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_COMPARE1_MASK);
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_COMPARE2_MASK);
-    nrf_rtc_int_disable(NRF_RTC2, NRF_RTC_INT_COMPARE3_MASK);
-    // For some reason even though the RTC should be stopped by now,
-    // the prescaler is still read-only. So, we loop here to make sure that the
-    // prescaler settings does take effect.
-    static const uint32_t prescaler = 4095;
-    while (rtc_prescaler_get(NRF_RTC2) != prescaler) {
-        nrf_rtc_prescaler_set(NRF_RTC2, prescaler);
-        __DSB();
-        __ISB();
-    }
-    // Start RTC2
-    nrf_rtc_task_trigger(NRF_RTC2, NRF_RTC_TASK_START);
-    __DSB();
-    __ISB();
-
-
-    Hal_Pin_Info* PIN_MAP = HAL_Pin_Map();
-
-    // Suspend all GPIOTE interrupts
-    HAL_Interrupts_Suspend();
-    {
-        uint32_t intenset = 0;
-        // Configure GPIOTE for wakeup
-        for (int i = 0; i < pins_count; i++) {
-            pin_t wake_up_pin = pins[i];
-            InterruptMode edge_trigger_mode = (i < mode_count) ? mode[i] : mode[mode_count - 1];
-
-            nrf_gpio_pin_pull_t wake_up_pin_mode;
-            nrf_gpio_pin_sense_t wake_up_pin_sense;
-            nrf_gpiote_polarity_t wake_up_pin_polarity;
-            // Set required pin_mode based on edge_trigger_mode
-            switch(edge_trigger_mode) {
-                case RISING: {
-                    wake_up_pin_mode = NRF_GPIO_PIN_PULLDOWN;
-                    wake_up_pin_sense = NRF_GPIO_PIN_SENSE_HIGH;
-                    wake_up_pin_polarity = NRF_GPIOTE_POLARITY_LOTOHI;
-                    break;
-                }
-                case FALLING: {
-                    wake_up_pin_mode = NRF_GPIO_PIN_PULLUP;
-                    wake_up_pin_sense = NRF_GPIO_PIN_SENSE_LOW;
-                    wake_up_pin_polarity = NRF_GPIOTE_POLARITY_HITOLO;
-                    break;
-                }
-                case CHANGE:
-                default:
-                    wake_up_pin_mode = NRF_GPIO_PIN_NOPULL;
-                    wake_up_pin_polarity = NRF_GPIOTE_POLARITY_TOGGLE;
-                    break;
-            }
-
-            // For any pin that is not currently configured in GPIOTE with IN event
-            // we are going to use low power PORT events
-            uint32_t nrf_pin = NRF_GPIO_PIN_MAP(PIN_MAP[wake_up_pin].gpio_port, PIN_MAP[wake_up_pin].gpio_pin);
-            // Set pin mode
-            nrf_gpio_cfg_input(nrf_pin, wake_up_pin_mode);
-            bool use_in = false;
-            for (unsigned i = 0; i < GPIOTE_CH_NUM; ++i) {
-                if (nrf_gpiote_event_pin_get(i) == nrf_pin &&
-                        nrf_gpiote_int_is_enabled(NRF_GPIOTE_INT_IN0_MASK << i)) {
-                    // We have to use IN event for this pin in order to successfully execute interrupt handler
-                    use_in = true;
-                    nrf_gpiote_event_configure(i, nrf_pin, wake_up_pin_polarity);
-                    intenset |= NRF_GPIOTE_INT_IN0_MASK << i;
-                    break;
-                }
-            }
-
-            if (!use_in) {
-                // Use PORT for this pin
-                if (wake_up_pin_mode == NRF_GPIO_PIN_NOPULL) {
-                    // Read current state, choose sense accordingly
-                    // Dummy read just in case
-                    (void)nrf_gpio_pin_read(nrf_pin);
-                    uint32_t cur_state = nrf_gpio_pin_read(nrf_pin);
-                    if (cur_state) {
-                        nrf_gpio_cfg_sense_set(nrf_pin, NRF_GPIO_PIN_SENSE_LOW);
-                    } else {
-                        nrf_gpio_cfg_sense_set(nrf_pin, NRF_GPIO_PIN_SENSE_HIGH);
-                    }
-                } else {
-                    nrf_gpio_cfg_sense_input(nrf_pin, wake_up_pin_mode, wake_up_pin_sense);
-                }
-                intenset |= NRF_GPIOTE_INT_PORT_MASK;
-            }
-        }
-
-        // Disable unnecessary GPIOTE interrupts
-        uint32_t cur_intenset = NRF_GPIOTE->INTENSET;
-        nrf_gpiote_int_disable(cur_intenset ^ intenset);
-        nrf_gpiote_int_enable(intenset);
-
-        // Clear events and interrupts
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_0);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_1);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_2);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_3);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_4);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_5);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_6);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_IN_7);
-        nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_PORT);
-        NVIC_ClearPendingIRQ(GPIOTE_IRQn);
-
-        if (pins_count > 0) {
-            exit_conditions |= STOP_MODE_EXIT_CONDITION_PIN;
-        }
-    }
-
-    int reason = SYSTEM_ERROR_UNKNOWN;
-
-    // Workaround for FPU anomaly
-    fpu_sleep_prepare();
-
-    // Masks all interrupts lower than softdevice. This allows us to be woken ONLY by softdevice
-    // or GPIOTE and RTC.
-    // IMPORTANT: No SoftDevice API calls are allowed until HAL_enable_irq()
-    int hst = 0;
-    bool advertising = hal_ble_gap_is_advertising(NULL); // We can now only restore advertising after exiting sleep mode.
-    if (!wakeup_by_ble) {
-        hal_ble_stack_deinit(NULL);
-        hst = HAL_disable_irq();
-    }
-
-    // Remember original priorities
-    uint32_t gpiote_priority = NVIC_GetPriority(GPIOTE_IRQn);
-    uint32_t rtc2_priority = NVIC_GetPriority(RTC2_IRQn);
-
-    // Enable RTC2 and GPIOTE interrupts if needed
-    if (exit_conditions & STOP_MODE_EXIT_CONDITION_PIN) {
-        NVIC_EnableIRQ(GPIOTE_IRQn);
-    }
-    if (seconds > 0) {
-        // Reconfigure RTC2 for wake-up
-        exit_conditions |= STOP_MODE_EXIT_CONDITION_RTC;
-        NVIC_ClearPendingIRQ(RTC2_IRQn);
-
-        nrf_rtc_event_clear(NRF_RTC2, NRF_RTC_EVENT_TICK);
-        nrf_rtc_event_enable(NRF_RTC2, RTC_EVTEN_TICK_Msk);
-        // Make sure that RTC is ticking
-        // See 'TASK and EVENT jitter/delay'
-        // http://infocenter.nordicsemi.com/index.jsp?topic=%2Fcom.nordic.infocenter.nrf52840.ps%2Frtc.html
-        while (!nrf_rtc_event_pending(NRF_RTC2, NRF_RTC_EVENT_TICK)) {
-            ;
-        }
-        nrf_rtc_event_disable(NRF_RTC2, RTC_EVTEN_TICK_Msk);
-        nrf_rtc_event_clear(NRF_RTC2, NRF_RTC_EVENT_TICK);
-
-        nrf_rtc_event_clear(NRF_RTC2, NRF_RTC_EVENT_COMPARE_0);
-        nrf_rtc_event_disable(NRF_RTC2, RTC_EVTEN_COMPARE0_Msk);
-
-        // Configure CC0
-        uint32_t counter = nrf_rtc_counter_get(NRF_RTC2);
-        uint32_t cc = counter + seconds * 8;
-        NVIC_EnableIRQ(RTC2_IRQn);
-        nrf_rtc_cc_set(NRF_RTC2, 0, cc);
-        nrf_rtc_event_clear(NRF_RTC2, NRF_RTC_EVENT_COMPARE_0);
-        nrf_rtc_int_enable(NRF_RTC2, NRF_RTC_INT_COMPARE0_MASK);
-        nrf_rtc_event_enable(NRF_RTC2, RTC_EVTEN_COMPARE0_Msk);
-    }
-
-    __DSB();
-    __ISB();
-
-    while (true) {
-        if ((exit_conditions & STOP_MODE_EXIT_CONDITION_RTC) && NVIC_GetPendingIRQ(RTC2_IRQn)) {
-            // Woken up by RTC
-            reason = 0;
-            break;
-        } else if ((exit_conditions & STOP_MODE_EXIT_CONDITION_PIN) && NVIC_GetPendingIRQ(GPIOTE_IRQn)) {
-            // Woken up by a pin, figure out which one
-            if (nrf_gpiote_event_is_set(NRF_GPIOTE_EVENTS_PORT)) {
-                // PORT event
-                for (int i = 0; i < pins_count; i++) {
-                    pin_t wake_up_pin = pins[i];
-                    uint32_t nrf_pin = NRF_GPIO_PIN_MAP(PIN_MAP[wake_up_pin].gpio_port, PIN_MAP[wake_up_pin].gpio_pin);
-                    nrf_gpio_pin_sense_t sense = nrf_gpio_pin_sense_get(nrf_pin);
-                    if (sense != NRF_GPIO_PIN_NOSENSE) {
-                        uint32_t state = nrf_gpio_pin_read(nrf_pin);
-                        if ((state && sense == NRF_GPIO_PIN_SENSE_HIGH) ||
-                                (!state && sense == NRF_GPIO_PIN_SENSE_LOW)) {
-                            reason = i + 1;
-                            break;
-                        }
-                    }
-                }
-            } else {
-                // Check IN events
-                for (unsigned i = 0; i < GPIOTE_CH_NUM && reason < 0; ++i) {
-                    if (NRF_GPIOTE->EVENTS_IN[i] && nrf_gpiote_int_is_enabled(NRF_GPIOTE_INT_IN0_MASK << i)) {
-                        pin_t pin = NRF_PIN_LOOKUP_TABLE[nrf_gpiote_event_pin_get(i)];
-                        if (pin != PIN_INVALID) {
-                            for (unsigned p = 0; p < pins_count; ++p) {
-                                pin_t wake_up_pin = pins[p];
-                                if (wake_up_pin == pin) {
-                                    reason = p + 1;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            break;
-        } else if (wakeup_by_ble && NVIC_GetPendingIRQ(SD_EVT_IRQn)) {
-            // FIXME: temporarily treat it as RTC wakeup reason.
-            reason = 0;
-            break;
-        }
-
-        {
-            // Mask interrupts completely
-            __disable_irq();
-            // Bump RTC2 and GPIOTE priorities if needed
-            if (exit_conditions & STOP_MODE_EXIT_CONDITION_PIN) {
-                NVIC_SetPriority(GPIOTE_IRQn, 0);
-            }
-            if (exit_conditions & STOP_MODE_EXIT_CONDITION_RTC) {
-                NVIC_SetPriority(RTC2_IRQn, 0);
-            }
-
-            __DSB();
-            __ISB();
-
-            // Go to sleep
-            __WFI();
-
-            // Unbump RTC2 and GPIOTE priorities
-            NVIC_SetPriority(GPIOTE_IRQn, gpiote_priority);
-            NVIC_SetPriority(RTC2_IRQn, rtc2_priority);
-
-            // Unmask interrupts, we are still under the effect of HAL_disable_irq
-            // that masked all but SoftDevice interrupts using BASEPRI
-            __enable_irq();
-        }
-    }
-
-    // Restore HFCLK
-    if (hfclk_resume) {
-        // Temporarily enable SoftDevice API interrupts
-        uint32_t base_pri = __get_BASEPRI();
-        __set_BASEPRI(_PRIO_SD_LOWEST << (8 - __NVIC_PRIO_BITS));
-        {
-            nrf_drv_clock_hfclk_request(NULL);
-            while (!nrf_drv_clock_hfclk_is_running()) {
-                ;
-            }
-        }
-        // And disable again
-        __set_BASEPRI(base_pri);
-    }
-
-
-    // Count the number of microseconds we've slept and reconfigure hal_timer (RTC2)
-    uint64_t slept_for = (uint64_t)nrf_rtc_counter_get(NRF_RTC2) * 125000;
-    // Stop RTC2
-    nrf_rtc_task_trigger(NRF_RTC2, NRF_RTC_TASK_STOP);
-    nrf_rtc_task_trigger(NRF_RTC2, NRF_RTC_TASK_CLEAR);
-    // Reconfigure it hal_timer_init() and apply the offset
-    hal_timer_init_config_t hal_timer_conf = {
-        .size = sizeof(hal_timer_init_config_t),
-        .version = 0,
-        .base_clock_offset = micros_before_sleep + slept_for
-    };
-    hal_timer_init(&hal_timer_conf);
-
-    // Restore GPIOTE cionfiguration
-    HAL_Interrupts_Restore();
-
-    // Re-initialize external flash
-    hal_exflash_init();
-
-    // Re-enable PWM
-    // FIXME: this is not ideal but will do for now
-    {
-        NRF_PWM_Type* const pwms[] = {NRF_PWM0, NRF_PWM1, NRF_PWM2, NRF_PWM3};
-        for (unsigned i = 0; i < sizeof(pwms) / sizeof(pwms[0]); i++) {
-            NRF_PWM_Type* pwm = pwms[i];
-            if (pwm_state[i]) {
-                nrf_pwm_enable(pwm);
-                if (nrf_pwm_event_check(pwm, NRF_PWM_EVENT_SEQEND0) ||
-                        nrf_pwm_event_check(pwm, NRF_PWM_EVENT_SEQEND1)) {
-                    nrf_pwm_event_clear(pwm, NRF_PWM_EVENT_SEQEND0);
-                    nrf_pwm_event_clear(pwm, NRF_PWM_EVENT_SEQEND1);
-                    if (pwm->SEQ[0].PTR && pwm->SEQ[0].CNT) {
-                        nrf_pwm_task_trigger(pwm, NRF_PWM_TASK_SEQSTART0);
-                    } else if (pwm->SEQ[1].PTR && pwm->SEQ[1].CNT) {
-                        nrf_pwm_task_trigger(pwm, NRF_PWM_TASK_SEQSTART1);
-                    }
-                }
-            }
-        }
-    }
-
-    // Re-enable SysTick
-    SysTick->CTRL |= SysTick_CTRL_ENABLE_Msk;
-
-    // This will reenable all non-SoftDevice interrupts previously disabled
-    // by sd_nvic_critical_region_enter()
-    sd_nvic_critical_region_exit(st);
-
-    // Unmasks all non-softdevice interrupts
-    if (!wakeup_by_ble) {
-        HAL_enable_irq(hst);
-        hal_ble_stack_init(NULL);
-        if (advertising) {
-            hal_ble_gap_start_advertising(NULL);
-        }
-    }
-
-    // Release LFCLK
-    nrf_drv_clock_lfclk_release();
-
-    // Re-enable USB
-    HAL_USB_Attach();
-
-    // Unlock external flash
-    hal_exflash_unlock();
-
-    // Enable thread scheduling
-    os_thread_scheduling(true, NULL);
-
-    return reason;
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
+    return SYSTEM_ERROR_DEPRECATED;
 }
 
 void HAL_Core_Execute_Stop_Mode(void) {
-    // Kept for compatibility only. Despite the fact that it's exported, it's never used directly
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
 }
 
 int HAL_Core_Enter_Standby_Mode(uint32_t seconds, uint32_t flags) {
-    // RTC cannot be kept running in System OFF mode, so wake up by RTC
-    // is not supported in deep sleep
-    if (seconds > 0) {
-        return SYSTEM_ERROR_NOT_SUPPORTED;
-    }
-
-    return HAL_Core_Execute_Standby_Mode_Ext(flags, NULL);
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
+    return SYSTEM_ERROR_DEPRECATED;
 }
 
 int HAL_Core_Execute_Standby_Mode_Ext(uint32_t flags, void* reserved) {
-    // Force to use external wakeup pin on Gen 3 Device
-    if (flags & HAL_STANDBY_MODE_FLAG_DISABLE_WKP_PIN) {
-        return SYSTEM_ERROR_NOT_SUPPORTED;
-    }
-
-    hal_pins_interrupt_config_t* pins_config = NULL;
-    const uint16_t* pins;
-    size_t pins_count;
-    const InterruptMode* modes;
-    size_t modes_count;
-
-    if (reserved) {
-        pins_config = (hal_pins_interrupt_config_t*)reserved;
-        pins = pins_config->pins;
-        pins_count = pins_config->pins_count;
-        modes = pins_config->modes;
-        modes_count = pins_config->modes_count;
-        if ((pins_count == 0 || modes_count == 0 || pins == NULL || modes == NULL)) {
-            pins_config = NULL;
-        }
-        for (unsigned i = 0; i < pins_count; i++) {
-            if (pins[i] >= TOTAL_PINS) {
-                pins_config = NULL;
-                break;
-            }
-        }
-        for (int i = 0; i < modes_count; i++) {
-            switch (modes[i]) {
-                case RISING:
-                case FALLING:
-                case CHANGE: {
-                    break;
-                }
-                default: {
-                    pins_config = NULL;
-                    break;
-                }
-            }
-        }
-    }
-
-    // Make sure we acquire exflash lock BEFORE going into a critical section
-    hal_exflash_lock();
-
-    // Disable thread scheduling
-    os_thread_scheduling(false, NULL);
-
-    // This will disable all but SoftDevice interrupts (by modifying NVIC->ICER)
-    uint8_t st = 0;
-    sd_nvic_critical_region_enter(&st);
-
-    // Disable SysTick
-    SysTick->CTRL &= ~SysTick_CTRL_ENABLE_Msk;
-
-    // Put external flash into sleep mode and disable QSPI peripheral
-    hal_exflash_special_command(HAL_EXFLASH_COMMAND_NONE, HAL_EXFLASH_COMMAND_SLEEP, NULL, NULL, 0);
-    hal_exflash_uninit();
-
-    // Uninit GPIOTE
-    nrfx_gpiote_uninit();
-
-    // Disable low power comparator
-    nrf_lpcomp_disable();
-
-    // Deconfigure any possible SENSE configuration
-    HAL_Interrupts_Suspend();
-
-    // Disable GPIOTE PORT interrupts
-    nrf_gpiote_int_disable(GPIOTE_INTENSET_PORT_Msk);
-
-    // Clear any GPIOTE events
-    nrf_gpiote_event_clear(NRF_GPIOTE_EVENTS_PORT);
-
-    Hal_Pin_Info* PIN_MAP = HAL_Pin_Map();
-
-    // Configure wakeup pin
-    if (!pins_config) {
-        uint32_t nrf_pin = NRF_GPIO_PIN_MAP(PIN_MAP[WKP].gpio_port, PIN_MAP[WKP].gpio_pin);
-        nrf_gpio_cfg_sense_input(nrf_pin, NRF_GPIO_PIN_PULLDOWN, NRF_GPIO_PIN_SENSE_HIGH);
-    } else {
-        for (int i = 0; i < pins_count; i++) {
-            pin_t wake_up_pin = pins[i];
-            InterruptMode edge_trigger_mode = (i < modes_count) ? modes[i] : modes[modes_count - 1];
-            nrf_gpio_pin_pull_t wake_up_pin_mode;
-            nrf_gpio_pin_sense_t wake_up_pin_sense;
-
-            // Set required pin_mode based on edge_trigger_mode
-            switch(edge_trigger_mode) {
-                case RISING: {
-                    wake_up_pin_mode = NRF_GPIO_PIN_PULLDOWN;
-                    wake_up_pin_sense = NRF_GPIO_PIN_SENSE_HIGH;
-                    break;
-                }
-                case FALLING: {
-                    wake_up_pin_mode = NRF_GPIO_PIN_PULLUP;
-                    wake_up_pin_sense = NRF_GPIO_PIN_SENSE_LOW;
-                    break;
-                }
-                case CHANGE:
-                default: {
-                    wake_up_pin_mode = NRF_GPIO_PIN_NOPULL;
-                    break;
-                }
-            }
-            uint32_t nrf_pin = NRF_GPIO_PIN_MAP(PIN_MAP[wake_up_pin].gpio_port, PIN_MAP[wake_up_pin].gpio_pin);
-            // Set pin mode
-            if (wake_up_pin_mode == NRF_GPIO_PIN_NOPULL) {
-                nrf_gpio_cfg_input(nrf_pin, wake_up_pin_mode);
-                // Read current state, choose sense accordingly
-                // Dummy read just in case
-                (void)nrf_gpio_pin_read(nrf_pin);
-                uint32_t cur_state = nrf_gpio_pin_read(nrf_pin);
-                if (cur_state) {
-                    nrf_gpio_cfg_sense_set(nrf_pin, NRF_GPIO_PIN_SENSE_LOW);
-                } else {
-                    nrf_gpio_cfg_sense_set(nrf_pin, NRF_GPIO_PIN_SENSE_HIGH);
-                }
-            } else {
-                nrf_gpio_cfg_sense_input(nrf_pin, wake_up_pin_mode, wake_up_pin_sense);
-            }
-        }
-    }
-
-    // Disable PWM
-    nrf_pwm_disable(NRF_PWM0);
-    nrf_pwm_disable(NRF_PWM1);
-    nrf_pwm_disable(NRF_PWM2);
-    nrf_pwm_disable(NRF_PWM3);
-
-    // RAM retention is configured on early boot in Set_System()
-
-    SPARK_ASSERT(sd_power_system_off() == NRF_SUCCESS);
-    while (1);
-    return 0;
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
+    return SYSTEM_ERROR_DEPRECATED;
 }
 
 void HAL_Core_Execute_Standby_Mode(void) {
+    // Deprecated. See: hal/src/nRF52840/sleep_hal.cpp
 }
 
 bool HAL_Core_System_Reset_FlagSet(RESET_TypeDef resetType) {
